Require equivalent VC build
Currently, this build pulls in whatever `yaml.dll` is in the environment's
Library\bin directory. This by default is the VS2008 / VC9 build, which
unncessarily introduces a dependency on VC9 which isn't directly covered
by Conda and breaks the expectation the the necessary DLLs for conda are
included within the environment.